Fix delegation in Ruby 3
Ruby 3 introduces separation between positional and keyword arguments
which breaks how this library handles delegation. The fix is to use the
argument forwarding shorthand (...) introduced in Ruby 2.7 to forward
all the arguments regardless if they're positional, keyword, or a block.

This does mean that backwards compatibility with Ruby 2.6 and older is
no longer supported. Given that Ruby 2.6 will reach EOL on 20220-03-31
I believe this is a fair compromise.